+++ src/tokenVerificationError.js
@@ -1,0 +1,15 @@
+/**
+ *
+ * @export
+ * @class TokenVerificationError
+ * @extends {Error}
+ */
+class TokenVerificationError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = 'TokenVerificationError';
+    this.code = 'Unauthorized';
+  }
+}
+
+module.exports = TokenVerificationError;